requirements.txt
Django==4.2.20
django-ckeditor-5==0.2.15
django-storages
django-db-logger
python-decouple
psycopg2-binary
boto3
channels
botocore
channels_redis
bs4
daphne
selenium
pandas
yfinance==0.2.54
celery
redis
django-redis
lxml
gunicorn
django-notifications-hq
